+++ src/test/java/CatTest.java
@@ -3,6 +3,7 @@ import com.example.Feline;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
+import org.mockito.Mockito;
 import org.mockito.junit.MockitoJUnitRunner;
 
 import java.util.List;
@@ -26,7 +27,8 @@ public class CatTest {
     @Test
     public void getFoodTest() throws Exception {
         Cat cat = new Cat(feline);
-        List<String> expected = List.of();
+        Mockito.when(cat.getFood()).thenReturn(List.of("Тест"));
+        List<String> expected = List.of("Тест");
         List<String> actual = cat.getFood();
         assertEquals(expected, actual);
     }
