+++ src/test/java/FelineTest.java
@@ -1,13 +1,10 @@
 import com.example.Feline;
 import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.junit.runners.Parameterized;
 
 import java.util.List;
 
 import static org.junit.Assert.assertEquals;
 
-@RunWith(Parameterized.class)
 public class FelineTest {
 
     @Test
@@ -31,32 +28,6 @@ public class FelineTest {
         Feline feline = new Feline();
         int expected = 1;
         int actual = feline.getKittens();
-        assertEquals(expected, actual);
-    }
-
-    // Parametrized test
-    private final int kittensNum;
-    private final int expected;
-
-    public FelineTest(int kittensNum, int expected) {
-        this.kittensNum = kittensNum;
-        this.expected = expected;
-    }
-
-    @Parameterized.Parameters
-    public static Object[][] getKittensNum() {
-        return new Object[][] {
-                { 10, 10},
-                { 0, 0},
-                { -1, -1},
-                { 1000000000, 1000000000},
-        };
-    }
-
-    @Test
-    public void getKittensWithOtherParams() {
-        Feline feline = new Feline();
-        int actual = feline.getKittens(kittensNum);
         assertEquals(expected, actual);
     }
 }
